+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VpWalletTest.kt
@@ -1,8 +1,9 @@
 package at.asitplus.wallet.lib.ktor.openid
 
 import at.asitplus.data.NonEmptyList.Companion.nonEmptyListOf
-import at.asitplus.dcapi.request.Oid4vpDCAPIRequest
+import at.asitplus.dcapi.request.DCAPIWalletRequest
 import at.asitplus.iso.IssuerSignedItem
+import at.asitplus.openid.AuthenticationResponseParameters
 import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.openid.OidcUserInfo
 import at.asitplus.openid.OidcUserInfoExtended
@@ -23,6 +24,7 @@ import at.asitplus.openid.dcql.DCQLQuery
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.testballoon.withFixtureGenerator
 import at.asitplus.wallet.eupid.EuPidScheme
+import at.asitplus.wallet.lib.RequestOptionsCredential
 import at.asitplus.wallet.lib.agent.ClaimToBeIssued
 import at.asitplus.wallet.lib.agent.CredentialToBeIssued
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithSelfSignedCert
@@ -38,6 +40,7 @@ import at.asitplus.wallet.lib.data.CredentialPresentation.DCQLPresentation
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest.DCQLRequest
 import at.asitplus.wallet.lib.data.SelectiveDisclosureItem
 import at.asitplus.wallet.lib.data.rfc3986.toUri
+import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.extensions.supportedSdAlgorithms
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception
 import at.asitplus.wallet.lib.oidvci.decodeFromPostBody
@@ -48,8 +51,7 @@ import at.asitplus.wallet.lib.openid.AuthnResponseResult.SuccessSdJwt
 import at.asitplus.wallet.lib.openid.ClientIdScheme
 import at.asitplus.wallet.lib.openid.OpenId4VpVerifier
 import at.asitplus.wallet.lib.openid.OpenId4VpVerifier.CreationOptions
-import at.asitplus.wallet.lib.openid.RequestOptions
-import at.asitplus.wallet.lib.openid.RequestOptionsCredential
+import at.asitplus.wallet.lib.openid.OpenId4VpRequestOptions
 import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
 import com.benasher44.uuid.uuid4
 import de.infix.testBalloon.framework.core.testSuite
@@ -95,7 +97,7 @@ val OpenId4VpWalletTest by testSuite {
                 clientId: String,
                 storeCredentials: Boolean = true,
             ) {
-                val requestOptions = RequestOptions(
+                val requestOptions = OpenId4VpRequestOptions(
                     credentials = setOf(
                         RequestOptionsCredential(
                             credentialScheme = scheme,
@@ -172,7 +174,7 @@ val OpenId4VpWalletTest by testSuite {
              */
             suspend fun setupRelyingPartyService(
                 clientId: String,
-                requestOptions: RequestOptions,
+                requestOptions: OpenId4VpRequestOptions,
                 validate: (AuthnResponseResult) -> Unit,
             ) {
                 val requestEndpointPath = "/request/${uuid4()}"
@@ -399,23 +401,25 @@ val OpenId4VpWalletTest by testSuite {
                        "response_type" : "vp_token"
                     }
                     """.trimIndent()
-            val dcApiRequest = Oid4vpDCAPIRequest(
-                protocol = Oid4vpDCAPIRequest.PROTOCOL_V1_UNSIGNED,
-                request = request,
+            val dcApiRequest = DCAPIWalletRequest.OpenId4VpUnsigned(
+                request = vckJsonSerializer.decodeFromString(request),
                 credentialId = "c72a2a8a6e94564cd8dea6ef0c7eb47b31a31947620ebcc0f07177bb71078def",
                 callingPackageName = "com.android.chrome",
                 callingOrigin = "https://apps.egiz.gv.at/customverifier"
             )
 
-            val input = joseCompliantSerializer.encodeToString(dcApiRequest)
-            val preparationState = it.wallet.startAuthorizationResponsePreparation(input).getOrThrow()
+            val preparationState = it.wallet.startAuthorizationResponsePreparation(dcApiRequest).getOrThrow()
             val presentation = DCQLPresentation(DCQLRequest(dcqlQuery), credentialQuerySubmissions)
             it.wallet.finalizeAuthorizationResponse(preparationState, presentation)
                 .getOrThrow()
                 .shouldBeInstanceOf<OpenId4VpWallet.AuthenticationForward>()
                 .authenticationResponseResult.shouldBeInstanceOf<AuthenticationResponseResult.DcApi>().apply {
-                    params.response shouldContain "vp_token"
-                    params.response shouldContain "cred1"
+                    val responseJson = vckJsonSerializer.encodeToString(
+                        AuthenticationResponseParameters.serializer(),
+                        params.data
+                    )
+                    responseJson shouldContain "\"vp_token\""
+                    responseJson shouldContain "\"cred1\""
                 }
         }