+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentIsoMdocTest.kt
@@ -274,7 +274,7 @@ private fun buildPresentationDefinition(vararg attributeName: String) = Presenta
                                 ).toString()
                             )
                         )
-                    }
+                    }.toSet()
                 )
             )
         ),